adding proper documentation to getModeratedChannels
