+++ app/src/main/java/com/example/clicker/network/clients/TwitchHomeClient.kt
@@ -1,20 +1,11 @@
 package com.example.clicker.network.clients
 
+import com.example.clicker.network.models.twitchClient.GetModChannels
 import com.example.clicker.network.models.twitchRepo.FollowedLiveStreams
-import com.google.gson.annotations.SerializedName
 import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.DELETE
-import retrofit2.http.Field
-import retrofit2.http.FormUrlEncoded
 import retrofit2.http.GET
 import retrofit2.http.Header
-import retrofit2.http.Headers
-import retrofit2.http.PATCH
-import retrofit2.http.POST
-import retrofit2.http.PUT
 import retrofit2.http.Query
-import javax.annotation.concurrent.Immutable
 
 
 /**
@@ -61,35 +52,6 @@ interface TwitchHomeClient {
     ):Response<GetModChannels>
 
 
-    /**
-     * - getAllFollowedStreamers represents a GET method. a function meant to get all of the user's channels they are following.
-     * This includes both live and not live channels
-     *
-     * @param authorization a String used to represent the OAuth token that uniquely identifies this user's granted abilities
-     * @param clientId a String used to represent the clientId(unique identifier) of this application
-     * @param userId a String used to represent the user who's followed channels we are requesting
-     *
-     * @return a [Response] object containing [AllFollowedStreamers]
-     * */
-    @GET("channels/followed")
-    suspend fun getAllFollowedStreamers(
-        @Header("Authorization") authorization: String,
-        @Header("Client-Id") clientId: String,
-        @Query("user_id") userId: String,
-        @Query("first") first: Int = 100
-    ): Response<AllFollowedStreamers>
+
 }
 
-
-@Immutable //Immutable added to make it stable with the list
-data class AllFollowedStreamers(
-    val total: Int,
-    val data: List<Streamer>
-)
-
-data class Streamer(
-    val broadcaster_id: String,
-    val broadcaster_login: String,
-    val broadcaster_name: String,
-    val followed_at: String
-)
